Move ANSI related code to separate module
